Renamed name property to title in IpatchSLI
IpatchSLI objects do not really have a name that can be changed so it
only returns the basename of the file name of the assigned file. The
only purpose is to have a name to display but title is better be used
for that so changed the name property to title now.