Fixed exporting multiple samples from instruments
Trying to export multiple samples from instruments was failing with a
crash. The reason for this were that the filename was determined from
the "name" property but only samples have that and not Zone or Region
objects that appear under instruments. Better to use the "title"
property instead which is supported by both.