Make Load Files dialog unique (#47)
This PR prevents the "Load Files" dialog to be opened multiple times.

We could think that it would be simpler to make the dialog modal, but the style "not modal" of the dialog is useful. It allows the user to load multiples files (using button apply) and come back in the main window to close files loaded by error keeping the dialog still opened to continue loading other files.