Load file unique (#48)
Actually swami allows loading of the same soundfont file multiple times. 
This has the following drawbacks:

Let loading the same file (f.sf2) 2 times, we can see 2 nodes in the Patches tab tree
(1)title
(2)title

- Now the user is making change on file (1) and then a different change on file (2).
- When the user save both files (using multiple selection and the Save Files dialog), as the 2 files  have the same path (f.sf2),  saving the second file (2) overwrite on the file system what was previously saved for the first file (1).

Not only the change made on file (1) are overwritten on the file system, but what is displayed in widget for this file (1) does not correspond to what was recorded on the file system. The user get quickly confused by an incoherent result.

To fix these issues, this PR avoid loading the same file multiple times. When loading a file, the loading function checks if the file is already loaded and if it is the case an information message is displayed and loading is skipped.