About file system "case sensitivity".
1)On linux file name on file system are "case sensitive".
2)On Windows file name is basically "case insensitive" but it could be
possible to set the file system "case sensitive".

But swami must ensure the following behaviour:
- prevent loading the same file multiple times (PR #48).
- prevent a "save file as" operation using a new name corresponding
  to a file in use by swami (this PR point 2.2.2 in the second part
  of first comment above).

Consequently to workaround the unpredictable file system behaviour, the
best for Swami is to consider file name "case insensitive" regardless
of the host OS. That means that f.sf2 should be considered the same file
as F.sf2.