GObject stopped accepting periods (`.`) in property names. (#59)
The `.` in GObject property names was causing a crash when I tried to open Preferences. This fixes it, at least for me.

This bug also causes `~/.config/swami/preferences.xml` to lose any of the FluidSynth settings with `.` in the name.

The issue started after I upgraded from Ubuntu 18.04 to 20.04.1, with GLib 2.64.3-1~ubuntu20.04.1. Presumably GObject changed the rules with the new version.

Not sure where the [GObject manual](https://developer.gnome.org/gobject/stable/) says that `.` isn't a valid part of a property name. Seems like that ought to be in there somewhere.

I probably missed a spot somewhere, but whatcha gonna do.